Turn TurtleLanguageServer into the actual TurtleLanguageServer
